# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fixit 24/7 API - On-demand home repair marketplace
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs/{jobId}/claims:
    get:
      operationId: listJobClaims
      tags: [claims]
      summary: List claims for a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Claims list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Claim"
    post:
      operationId: claimJob
      tags: [claims]
      summary: Claim a job (tradie)
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClaimBody"
      responses:
        "201":
          description: Job claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
        "409":
          description: Claim limit reached or already claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{jobId}/claims/{claimId}:
    put:
      operationId: updateClaim
      tags: [claims]
      summary: Update claim status (homeowner accepts/rejects, tradie withdraws)
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
        - name: claimId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClaimBody"
      responses:
        "200":
          description: Updated claim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
    delete:
      operationId: withdrawClaim
      tags: [claims]
      summary: Withdraw a claim (tradie)
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
        - name: claimId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawn
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Notifications
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Claim:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        tradieId:
          type: integer
        tradieName:
          type: string
          nullable: true
        tradieRating:
          type: number
          nullable: true
        tradieReviewCount:
          type: integer
          nullable: true
        tradieSuburb:
          type: string
          nullable: true
        tradieAvatarUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, accepted, rejected, withdrawn, completed]
        message:
          type: string
          nullable: true
        proposedPrice:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - jobId
        - tradieId
        - status
        - createdAt
        - tradieReviewCount
    CreateClaimBody:
      type: object
      properties:
        message:
          type: string
        proposedPrice:
          type: number
    UpdateClaimBody:
      type: object
      properties:
        status:
          type: string
          enum: [accepted, rejected, withdrawn, completed]
      required:
        - status
...